requirements.txt
elasticsearch==8.8.0
sentence-transformers>=2.2.2
openai>=1.7.2
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
datasets>=2.14.0
accelerate>=1.1.0
